Only remove old storage directory if MEDIA_ROOT is set, and exists.

+++ src/maasserver/migrations/0041_remove_filestorage_data.py
@@ -21,9 +21,10 @@ class Migration(SchemaMigration):
         db.delete_column(u'maasserver_filestorage', 'data')
 
         # Cleaning up any obsolete FileStorage files.
-        rmtree(
-            os.path.join(settings.MEDIA_ROOT, upload_dir),
-            ignore_errors=True)
+        if settings.MEDIA_ROOT and os.path.isdir(settings.MEDIA_ROOT):
+            rmtree(
+                os.path.join(settings.MEDIA_ROOT, upload_dir),
+                ignore_errors=True)
 
     def backwards(self, orm):
